Factor out test setup and teardown
Setting up and tearing down test resources is now done by a shared function.
