+++ e2e/node/resource.test.ts
@@ -54,6 +54,7 @@ import {
 } from "../util/getTestingEnvironment";
 import { getAuthenticatedSession } from "../util/getAuthenticatedSession";
 import type { Session } from "@inrupt/solid-client-authn-node";
+import { setupTestResources, teardownTestResources } from "./test-setup";
 
 const env: TestingEnvironment = getTestingEnvironment();
 
@@ -61,6 +62,8 @@ if (env.environment === "NSS") {
   // eslint-disable-next-line jest/no-focused-tests
   test.only(`Skipping Unauth NSS tests in ${env.environment}`, () => {});
 }
+
+const TEST_SLUG = "solid-client-test-e2e-resource"
 
 describe("Authenticated end-to-end", () => {
   let options: { fetch: typeof global.fetch };
@@ -70,38 +73,14 @@ describe("Authenticated end-to-end", () => {
 
   beforeEach(async () => {
     session = await getAuthenticatedSession(env);
-    // Set the user agent to something distinctive to make debug easier
-    const fetchWithAgent = (url: RequestInfo, options?: RequestInit) => {
-      return session.fetch(url, {
-        ...options,
-        headers: {
-          ...options?.headers,
-          "User-Agent": "solid-client node e2e tests"
-        }
-      })
-    }
+    const { containerUrl, resourceUrl, fetchWithAgent } = await setupTestResources(session, TEST_SLUG, env.pod);
+    sessionResource = resourceUrl;
+    sessionContainer = containerUrl;
     options = { fetch: fetchWithAgent };
-    sessionContainer = getSourceIri(
-      await createContainerInContainer(
-        env.pod, {
-          ...options,
-          slugSuggestion: "solid-client-tests-node-e2e",
-        }
-      )
-    );
-    sessionResource = getSourceIri(
-      await saveSolidDatasetInContainer(
-        sessionContainer,
-        createSolidDataset(),
-        options
-      )
-    );
   });
 
   afterEach(async () => {
-    await deleteSolidDataset(sessionResource, options);
-    await deleteSolidDataset(sessionContainer, options);
-    await session.logout();
+    await teardownTestResources(session, sessionContainer, sessionResource, options.fetch);
   });
 
   it("can create, read, update and delete data", async () => {
